requirements.txt
flask==2.0.1
psycopg2-binary==2.9.1
tabulate==0.8.9
gunicorn==20.1.0
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
alembic==1.12.1
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
